Add all scopes as optional
Why:

For the [node hubspot api client], I would like to have an example to
show how to obtain an auth token with the appropriate scopes for running
the test suite with mocking disabled.

[node hubspot api client]: https://github.com/MadKudu/node-hubspot

This PR:

Adds all the hubspot scopes as "Optional" meaning that they'll only be
asked for access if those features are enabled.